requirements.txt
discord.py==2.6.2
httpx==0.28.1